% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISTScore.R
\name{CMISTScore}
\alias{CMISTScore}
\title{Calculate the CMIST Score}
\usage{
CMISTScore(risks, uncertainties)
}
\arguments{
\item{risks}{a vector containing the risk scores of the 17 CMIST questions}

\item{uncertainties}{a vector containing the uncertainty scores of the 17 CMIST questions}
}
\value{

}
\description{
Calculate the CMIST Score
}
\examples{
set.seed(11)
risks <- sample(x = c(1:3),size = 17,replace = TRUE)
uncertainties <- sample(x = c(1:3),size = 17,replace = TRUE)
score <- CMISTScore(risks,uncertainties)
}
