% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CMIST_database.R
\name{get_CMIST_database}
\alias{get_CMIST_database}
\title{Get the CMIST data from the online portal}
\usage{
get_CMIST_database()
}
\value{

}
\description{
Get the CMIST data from the online portal
}
\examples{
CMISTdata <- get_CMIST_database()
}
