% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probs}
\alias{probs}
\title{The probability of each simulated score.}
\format{
A data frame with 27 rows and 4 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{https://link.springer.com/article/10.1007/s10530-015-1008-y#Sec10}
}
\usage{
probs
}
\description{
The probability of each simulated score based on the risk and probability.
Sourced from Supplementary material 2 in:
Drolet, D., DiBacco, C., Locke, A. et al. Evaluation of a new screening-level
risk assessment tool applied to non-indigenous marine invertebrates in
Canadian coastal waters. Biol Invasions 18, 279–294 (2016).
https://doi.org/10.1007/s10530-015-1008-y
}
\keyword{datasets}
