% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simScore.R
\name{simScore}
\alias{simScore}
\title{Simulate scores based on risk and uncertainty}
\usage{
simScore(risk, certainty, n = 1000)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Simulate scores based on risk and uncertainty
}
\examples{
scores <- simScore(risk=2,certainty=1)
}
