% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISTScore.R
\name{CMISTScore}
\alias{CMISTScore}
\title{Title}
\usage{
CMISTScore(risks, uncertainties)
}
\arguments{
\item{uncertainties}{}
}
\value{

}
\description{
Title
}
\examples{
set.seed(11)
risks <- sample(x = c(1:3),size = 17,replace = TRUE)
uncertainties <- sample(x = c(1:3),size = 17,replace = TRUE)
score <- CMISTScore(risks,uncertainties)
}
